+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [8.0.0] - 2022-05-25
 
+### Added
+- Add support for Show Resource endpoint
+  [conjur-api-python#31](https://github.com/cyberark/conjur-api-python/pull/31)
+- Add support for LDAP authentication
+  [conjur-api-python#22](https://github.com/cyberark/conjur-api-python/pull/22)
+
 ### Changed
 - Abstract authentication flow into new `AuthenticationStrategyInterface`
   [conjur-api-python#20](https://github.com/cyberark/conjur-api-python/pull/20)
-- Add support for LDAP authentication
-  [conjur-api-python#22](https://github.com/cyberark/conjur-api-python/pull/22)
 - Store API key in `CreditentialsData` object
   [conjur-api-python#23](https://github.com/cyberark/conjur-api-python/pull/23)